+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 0.0.2
 
 * Updated LICENSE
+
+## 0.0.4
+
+* Refactor from JAVA to Kotlin
+* Update plugin register rules